refactor(auth): removes authN logic from Core, as that moved away to the manager
+ modifies tests to allow for this change
