+++ Core.sln
@@ -32,9 +32,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Client.Serializer.S
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphQL.Client.LocalExecution", "GraphQLClient\src\GraphQL.Client.LocalExecution\GraphQL.Client.LocalExecution.csproj", "{6A3861F4-CA3E-4790-BF27-3C502E18783C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServerTransport", "ServerTransport\ServerTransport.csproj", "{899C8BE7-C206-4698-8807-3693FDF8A3A2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServerTransport", "ServerTransport\ServerTransport.csproj", "{899C8BE7-C206-4698-8807-3693FDF8A3A2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MongoDBTransport", "MongoDBTransport\MongoDBTransport.csproj", "{4450B0E9-D663-4F13-AF96-6A41B1A3B5C9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MongoDBTransport", "MongoDBTransport\MongoDBTransport.csproj", "{4450B0E9-D663-4F13-AF96-6A41B1A3B5C9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Transports", "Transports", "{424ECC59-E693-4678-AFAC-C00F2D8AE735}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{A3E889EB-76D7-4A6E-A3FB-BCAEAC2634FC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -86,11 +90,21 @@ Global
 		{6A3861F4-CA3E-4790-BF27-3C502E18783C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6A3861F4-CA3E-4790-BF27-3C502E18783C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6A3861F4-CA3E-4790-BF27-3C502E18783C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{899C8BE7-C206-4698-8807-3693FDF8A3A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{899C8BE7-C206-4698-8807-3693FDF8A3A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{899C8BE7-C206-4698-8807-3693FDF8A3A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{899C8BE7-C206-4698-8807-3693FDF8A3A2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4450B0E9-D663-4F13-AF96-6A41B1A3B5C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4450B0E9-D663-4F13-AF96-6A41B1A3B5C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4450B0E9-D663-4F13-AF96-6A41B1A3B5C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4450B0E9-D663-4F13-AF96-6A41B1A3B5C9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
+		{2EDE05B7-2B21-4C0B-9EA5-3BC7A52315BE} = {A3E889EB-76D7-4A6E-A3FB-BCAEAC2634FC}
+		{4A79F5D2-3B66-43E3-BE33-0BB79FBFDC61} = {A3E889EB-76D7-4A6E-A3FB-BCAEAC2634FC}
 		{390B0BA1-4621-459F-8CE2-788797AB3134} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
 		{A71D3F8F-4592-4CD4-83E4-A215D116C730} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
 		{87A3348F-A510-4101-BD87-DBBF9EBE4070} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
@@ -98,6 +112,8 @@ Global
 		{54B5BF04-3C46-498C-AE2A-1F2F85768C2C} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
 		{EAC54DBF-B8FC-4B20-A33A-3EBA05841128} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
 		{6A3861F4-CA3E-4790-BF27-3C502E18783C} = {1099033B-8BD9-4038-AF9A-94A3370B4F8C}
+		{899C8BE7-C206-4698-8807-3693FDF8A3A2} = {424ECC59-E693-4678-AFAC-C00F2D8AE735}
+		{4450B0E9-D663-4F13-AF96-6A41B1A3B5C9} = {424ECC59-E693-4678-AFAC-C00F2D8AE735}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1378F8F4-B354-4DB6-8052-A1A69B59BC2C}
