+++ Core/Credentials/Account.cs
@@ -58,48 +58,6 @@ namespace Speckle.Core.Credentials
       return response.Data.user;
     }
 
-    public async Task RotateToken()
-    {
-      using (var client = new HttpClient())
-      {
-        var request = new HttpRequestMessage()
-        {
-          RequestUri = new Uri(new Uri(serverInfo.url), "auth/token"),
-          Method = HttpMethod.Post
-        };
-
-        var payload = new
-        {
-          appId = AccountManager.APPID,
-          appSecret = AccountManager.SECRET,
-          refreshToken
-        };
-
-        request.Content = new StringContent(Newtonsoft.Json.JsonConvert.SerializeObject(payload));
-
-        request.Content.Headers.ContentType = MediaTypeHeaderValue.Parse("application/json");
-        request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-
-        var _response = await client.SendAsync(request);
-
-        try
-        {
-          _response.EnsureSuccessStatusCode();
-        }
-        catch (Exception e)
-        {
-          Log.CaptureAndThrow(new SpeckleException($"Failed to rotate token. Response status: {_response.StatusCode}.", e));
-        }
-
-        var _tokens = JsonConvert.DeserializeObject<TokenExchangeResponse>(await _response.Content.ReadAsStringAsync());
-
-        refreshToken = _tokens.refreshToken;
-        token = _tokens.token;
-
-        AccountManager.UpdateOrSaveAccount(this);
-      }
-    }
-
     public bool Equals(Account other)
     {
       return other.userInfo.email == userInfo.email && other.serverInfo.url == serverInfo.url;
