+++ IntegrationTests/Utils.cs
@@ -5,9 +5,26 @@ using System.Text;
 using Newtonsoft.Json;
 using Speckle.Core.Api;
 using Speckle.Core.Credentials;
+using Speckle.Core.Transports;
 
 namespace TestsIntegration
 {
+  public class Fixtures
+  {
+    private static SQLiteTransport AccountStorage = new SQLiteTransport(scope: "Accounts");
+
+    public static void UpdateOrSaveAccount(Account account)
+    {
+      AccountStorage.DeleteObject(account.id);
+      AccountStorage.SaveObjectSync(account.id, JsonConvert.SerializeObject(account));
+    }
+
+    public static void DeleteLocalAccount(string id)
+    {
+      AccountStorage.DeleteObject(id);
+    }
+  }
+
   public static class Utils
   {
     public static Account SeedUser(ServerInfo server)
